requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
bcrypt==4.1.3
httpx==0.27.0
APScheduler==3.10.4
openpyxl==3.1.2
pytz==2024.1
